grilla de incidentes, arreglo conexion db

+++ CALL CENTER/dominio/Incidente.cs
@@ -9,13 +9,13 @@ namespace dominio
     public class Incidente
     {
         public string Cliente { get; set; }
-        public int EmpleadoLegajo { get; set; }
+        public long EmpleadoLegajo { get; set; }
         public string Tipo { get; set; }
         public string Prioridad { get; set; }
         public string Asunto { get; set; }
         public string Estado { get; set; }
         public string Email { get; set; }
-        public int ID { get; set; }
+        public long ID { get; set; }
         public string ComentarioCierre { get; set; }
     }   
 }
